+++ packages/at_chops/lib/src/at_chops_base.dart
@@ -1,14 +1,9 @@
 import 'dart:typed_data';
 
+import 'package:at_chops/at_chops.dart';
 import 'package:at_chops/src/algorithm/at_algorithm.dart';
-import 'package:at_chops/src/algorithm/at_iv.dart';
-import 'package:at_chops/src/algorithm/rsa_encryption_algo.dart';
 import 'package:at_chops/src/algorithm/default_hashing_algo.dart';
-import 'package:at_chops/src/key/impl/at_chops_keys.dart';
-import 'package:at_chops/src/key/key_type.dart';
-import 'package:at_chops/src/metadata/at_signing_input.dart';
-import 'package:at_chops/src/metadata/encryption_result.dart';
-import 'package:at_chops/src/metadata/signing_result.dart';
+import 'package:at_chops/src/factory/at_hashing_algo_factory.dart';
 
 /// Base class for all Cryptographic and Hashing Operations. Callers have to either implement
 /// specific encryption, signing or hashing algorithms. Otherwise default implementation of specific algorithms will be used.
@@ -18,6 +13,25 @@ abstract class AtChops {
   AtChopsKeys get atChopsKeys => _atChopsKeys;
 
   AtChops(this._atChopsKeys);
+
+  /// Returns an instance of [AtHashingAlgorithm] based on the provided [hashingAlgoType].
+  ///
+  /// This static method acts as a factory to create a hashing algorithm object, which
+  /// can be used to hash data using the specified hashing algorithm type.
+  ///
+  /// The [hashingAlgoType] parameter defines the type of hashing algorithm to be used.
+  /// It should be an instance of [HashingAlgoType], which represents different supported
+  /// hashing algorithms like SHA512, MD5, etc.
+  ///
+  /// Example usage:
+  /// ```dart
+  /// String hashedValue = AtChops.hashWith(HashingAlgoType.sha512).hash('some-text');
+  /// ```
+  ///
+  /// [hashingAlgoType]: The type of the hashing algorithm (e.g., SHA256, MD5).
+  /// Returns an [AtHashingAlgorithm] instance based on the provided [hashingAlgoType].
+  static AtHashingAlgorithm hashWith(HashingAlgoType hashingAlgoType) =>
+      AtHashingAlgorithmFactory.withHashingAlgorithm(hashingAlgoType);
 
   /// Encrypts the input bytes [data] using an [encryptionAlgorithm] and returns [AtEncryptionResult].
   /// If [encryptionKeyType] is [EncryptionKeyType.rsa2048] then [encryptionAlgorithm] will be set to [RsaEncryptionAlgo]