+++ packages/at_commons/lib/src/keystore/public_key_hash.dart
@@ -1,19 +1,26 @@
+import 'dart:convert';
+
+import 'package:at_commons/at_commons.dart';
+
 /// Represents hash of an atsign's public encryption key and the hashing algorithm used
 class PublicKeyHash {
+  // Holds the hashed value of the data.
   String hash;
-  PublicKeyHashingAlgo publicKeyHashingAlgo;
 
-  PublicKeyHash(this.hash, this.publicKeyHashingAlgo);
+  // Holds the algorithm used to the hash the data.
+  String hashingAlgo;
+
+  PublicKeyHash(this.hash, this.hashingAlgo);
 
   @override
   String toString() {
-    return 'PublicKeyHash{hash: $hash, publicKeyHashingAlgo: $publicKeyHashingAlgo}';
+    return jsonEncode(toJson());
   }
 
   Map toJson() {
     var map = {};
-    map['hash'] = hash;
-    map['algo'] = publicKeyHashingAlgo.name;
+    map[AtConstants.sharedWithPublicKeyHashValue] = hash;
+    map[AtConstants.sharedWithPublicKeyHashingAlgo] = hashingAlgo;
     return map;
   }
 
@@ -21,8 +28,8 @@ class PublicKeyHash {
     if (json == null) {
       return null;
     }
-    return PublicKeyHash(
-        json['hash'], PublicKeyHashingAlgo.values.byName(json['algo']));
+    return PublicKeyHash(json[AtConstants.sharedWithPublicKeyHashValue],
+        json[AtConstants.sharedWithPublicKeyHashingAlgo]);
   }
 
   @override
@@ -31,10 +38,8 @@ class PublicKeyHash {
       other is PublicKeyHash &&
           runtimeType == other.runtimeType &&
           hash == other.hash &&
-          publicKeyHashingAlgo == other.publicKeyHashingAlgo;
+          hashingAlgo == other.hashingAlgo;
 
   @override
-  int get hashCode => hash.hashCode ^ publicKeyHashingAlgo.hashCode;
+  int get hashCode => hash.hashCode ^ hashingAlgo.hashCode;
 }
-
-enum PublicKeyHashingAlgo { sha256, sha512 }