+++ tests/at_onboarding_cli_functional_tests/test/at_onboarding_cli_test.dart
@@ -147,12 +147,23 @@ void main() {
           AtOnboardingServiceImpl(atSign, atOnboardingPreference);
       bool status = await atOnboardingService.onboard();
       expect(status, true);
+      // verify whether ttr is set in public encryption key
+      final atLookup = AtLookupImpl(atSign, atOnboardingPreference.rootDomain,
+          atOnboardingPreference.rootPort);
+      var encryptionPublicKey =
+          await atLookup.executeCommand('lookup:all:publickey$atSign\n');
+      expect(encryptionPublicKey, isNotNull);
+      encryptionPublicKey = encryptionPublicKey?.replaceFirst('data:', '');
+      expect(jsonDecode(encryptionPublicKey!)['metaData'], isNotNull);
+      expect(jsonDecode(encryptionPublicKey)['metaData']['ttr'], -1);
+      print('encryptionPublicKey: $encryptionPublicKey');
       bool status2 = await atOnboardingService.authenticate();
       expect(status2, true);
       expect(await atOnboardingService.isOnboarded(), true);
 
       /// Assert .atKeys file is generated for the atSign
       expect(await File(atOnboardingPreference.atKeysFilePath!).exists(), true);
+      await atLookup.close();
     });
 
     tearDown(() async {